# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hotel Booking API
  version: 0.1.0
  description: |
    Backend HTTP API for the Hotel Booking SaaS. Powers the booking-web
    (guest-facing) and admin-web (operator-facing) Next.js apps.

    All authenticated endpoints expect a JWT access token in the
    `Authorization: Bearer <token>` header. Error responses use a shared
    envelope: `{ "error": { "code": "STRING", "message": "STRING" } }`.

    Money fields ending in `_cents` are integers in the currency's minor unit
    (e.g. THB satang); decimal-string fields (`rate`, `amount`, etc.) are
    used when JS precision matters.
  contact:
    name: Hotel Booking Engineering
    email: engineering@example.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /v1/hotels:
    get:
      tags: [hotels]
      summary: List hotels owned by the caller's account
      responses:
        '200':
          description: Hotel list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HotelListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [hotels]
      summary: Create a hotel (owner / manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HotelCreateRequest' }
      responses:
        '201':
          description: Hotel created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Hotel' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/slug-available:
    get:
      tags: [hotels]
      summary: Check whether a public slug is unused
      parameters:
        - in: query
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Availability flag
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SlugAvailableResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{id}:
    parameters:
      - $ref: '#/components/parameters/HotelIdPath'
    get:
      tags: [hotels]
      summary: Get a hotel by id
      responses:
        '200':
          description: Hotel
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Hotel' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      tags: [hotels]
      summary: Partial update (owner / manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HotelUpdateRequest' }
      responses:
        '200':
          description: Updated hotel
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Hotel' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [hotels]
      summary: Soft-delete a hotel (owner only)
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ----- room types -----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /v1/auth/login or /v1/auth/refresh.
  parameters:
    HotelIdPath:
      in: path
      name: id
      required: true
      description: Hotel id
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid credentials / token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Caller's role does not permit this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found (or hidden by tenant isolation)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State / uniqueness conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine code (e.g. NOT_FOUND, BAD_REQUEST)
            message: { type: string }

    # ----- auth -----
...
    Hotel:
      type: object
      required: [id, account_id, slug, name, timezone, base_currency, check_in_time, check_out_time, kyc_status, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        hotel_type: { type: string }
        description: { type: string }
        address_line: { type: string }
        city: { type: string }
        country: { type: string }
        postal_code: { type: string }
        latitude: { type: [number, 'null'], format: double }
        longitude: { type: [number, 'null'], format: double }
        phone: { type: string }
        email: { type: string, format: email }
        line_id: { type: string }
        timezone: { type: string }
        base_currency: { type: string, description: ISO 4217 code }
        check_in_time: { type: string, description: 'HH:MM:SS' }
        check_out_time: { type: string, description: 'HH:MM:SS' }
        kyc_status:
          type: string
          enum: [pending, submitted, approved, rejected]
        status:
          type: string
          enum: [test, live, suspended, archived]
        promptpay_id:
          type: [string, 'null']
          description: |
            Optional PromptPay receiver. One of: 10-digit phone (e.g.
            "0812345678"), 13-digit Thai national ID, or 15-char e-Wallet /
            tax-ID. Dashes/spaces are stripped before validation.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    HotelCreateRequest:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
        hotel_type: { type: string }
        country: { type: string }
        timezone: { type: string }
        base_currency: { type: string }
        promptpay_id:
          type: [string, 'null']
          description: 10-digit phone, 13-digit national ID, or 15-char tax ID.
    HotelUpdateRequest:
      type: object
      description: |
        Partial PATCH; only non-null fields are written. Slug is intentionally
        not updatable here. `promptpay_id` is tri-state: omit to leave
        unchanged, set to an empty string to clear, or send a non-empty
        value to set.
      properties:
        name: { type: [string, 'null'] }
        hotel_type: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        address_line: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        country: { type: [string, 'null'] }
        postal_code: { type: [string, 'null'] }
        latitude: { type: [number, 'null'], format: double }
        longitude: { type: [number, 'null'], format: double }
        phone: { type: [string, 'null'] }
        email: { type: [string, 'null'], format: email }
        line_id: { type: [string, 'null'] }
        timezone: { type: [string, 'null'] }
        base_currency: { type: [string, 'null'] }
        check_in_time: { type: [string, 'null'] }
        check_out_time: { type: [string, 'null'] }
        promptpay_id: { type: [string, 'null'] }
    HotelListResponse:
      type: object
      required: [hotels]
      properties:
        hotels:
          type: array
          items: { $ref: '#/components/schemas/Hotel' }
    SlugAvailableResponse:
      type: object
      required: [slug, available]
      properties:
        slug: { type: string }
        available: { type: boolean }

    # ----- room type -----
...
security:
  - bearerAuth: []